requirements.txt
APScheduler==3.6.3
async-lru==1.0.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
confuse==1.6.0
Jinja2==3.0.2
MarkupSafe==2.0.1
pip==21.3.1
pyaes==1.6.1
Pyrogram==1.2.9
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
setuptools==41.2.0
six==1.16.0
TgCrypto==1.2.2
tornado==6.1
tzdata==2021.5
tzlocal==4.0.2
